Implement firmware to control Makeblock XY-plotter
* The plotter ships with an Arduino compatible controller board which may be used as a reference when we
evaluate the outcome of the project.

* The firmware receives and interprets G-codes and driver the mechanics accordingly. G-codes are fairly
simple instructions that tell for example the plotter to move to a certain point or to draw an arc on the
canvas. Some G codesread/write settings of the device.

* The plotter consists of frame, two stepper motors and their driver chips, a servo to control pencil
movement and a PWM driven laser engraver.

* The plotter is controlled with mDraw. mDraw can be downloaded from Makeblock web site. Your
LPCXpresso connects to the computer through a USB-serial port (target port). The other USB serial port that
is provided by the debugger is to be used only for debugging.

* 2 RJ-11 connectors for driving the stepper motors. Each connector is driven by two IO-pins:
direction and step. The details of the driver chip can be found in that datasheet in the workspace.

* 3 RJ-11 connectors for reading limit switches and for controlling the pencil.

* 1 screw terminal for controlling the laser engraver.

* The pencil is mounted to a servo that controls pencil’s vertical
  position. The servo is controlled by PWM.

* Laser engraver beam intensity is controlled by pulsing the input power of the laser module. The laser
module can draw up to 500mA at 12V thus a transistor driver is provided to driver the laser. The transistor
driver is also driven by PWM.